Vehicles ID
Now every vehicle has an ID assigned to it. When a vehicle is removed, its ID will become available to be used by a new vehicle (when a new vehicle is added, the lowest ID available will be assigned to it).
Also, discoverd (and solved) some issues with the removal of selected cars.