Rent and return
Managed to make the rent function for sedans. Also, made the option to return sedans (right now it's pretty basic, it just adds a new sedan with the old id, name, fuel and damage ). I'm planning to do another UserControl that will prompt the user to enter the fuel and damage percentages at return, and also to specify the date it has been returned for determining whether the vehicle has been returned on time or late.